Auto defense moves with raise/low to get over high wall.

+++ Stronghold/src/Subsystems/Shooter.h
@@ -86,7 +86,7 @@ public:
 	void ShooterStop();
 	bool ShooterIsHome();
 	bool ShooterAtMax();
-	float GetShooterElevation();
+	int GetShooterElevation();
 
 	// Hook Manipulation public functions
 	void HookRaise();
